requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3